* Ported filters to new architecture:   Brightness, ColorBoost, ColorEnhancement, ColorIntensity, LabTransform, DRC, Levels, LMHRecovery,   Normalization, ReinhardBrighten, Saturation, Tone * Added FUID constants for every filter * Some double -> float in ptImage

+++ Sources/ptChannelMixer.h
@@ -35,7 +35,7 @@
 class ptChannelMixer {
 public :
 
-double m_Mixer[3][3]; // [To][From]
+float m_Mixer[3][3]; // [To][From]
 
 // Reading and writing function (compatible).
 // Header is a free text that is inserted as comment to describe
